Modify lval for sexpr & error strings
Lval will now hold sexpressions and error strings. This means that
pointers will be present in the struct. This commit updates the
constructor functions, the struct definition, adds constructors for
new types, and adds a destructor.

It is not in a stable state as the language handling hasn't been
attached to the new lval type.